Introduce a new concourse-secrets-admin module
This is part of work addressing the situation regarding teams
accessing Concourse secrets, providing a mechanism to manage secrets
without having to hijack/intercept Concourse containers.

This commit adds a module for creating a IAM role and associated
policy, with the aim that future work will allow assuming that role
through GitHub authentication and membership of a specific GitHub
team.

https://trello.com/c/v6tDw4Yo/1310-set-up-terraform-roles-etc-for-cognoto